kymotrack: renamed module from `kymoline`
