+++ lumicks/pylake/kymotracker/tests/test_stitching.py
@@ -1,7 +1,7 @@
 import numpy as np
 from lumicks.pylake.kymotracker.detail.stitch import distance_line_to_point
 from lumicks.pylake.kymotracker.stitching import stitch_kymo_lines
-from lumicks.pylake.kymotracker.kymoline import KymoTrack
+from lumicks.pylake.kymotracker.kymotrack import KymoTrack
 
 
 def test_distance_line_to_point():
