+++ lumicks/pylake/tests/test_imaging_confocal/conftest.py
@@ -272,7 +272,7 @@ def kymo_h5_file(tmpdir_factory, reference_counts):
 
     # Force channel that overlaps kymo; step from high to low force
     # We want two lines of the kymo to have a force of 30, the other 10. Force starts 5 samples
-    # before the kymograph. First kymoline is 15 samples long, second is 16 samples long, which
+    # before the kymograph. First kymotrack is 15 samples long, second is 16 samples long, which
     # means the third line starts after 31 + 5 = 36 samples
     force_data = np.hstack((np.ones(37) * 30, np.ones(33) * 10))
     force_start = np.int64(ds.attrs["Start time (ns)"] - (dt * 5))  # before infowave
